+++ kubernetes-model-generator/kubernetes-model-jsonschema2pojo/src/main/java/io/fabric8/kubernetes/jsonschema2pojo/KubernetesTypeAnnotator.java
@@ -15,43 +15,60 @@
  */
 package io.fabric8.kubernetes.jsonschema2pojo;
 
-import com.sun.codemodel.*;
+import com.sun.codemodel.JAnnotationArrayMember;
+import com.sun.codemodel.JAnnotationUse;
+import com.sun.codemodel.JClassAlreadyExistsException;
+import com.sun.codemodel.JCodeModel;
+import com.sun.codemodel.JDefinedClass;
 import io.sundr.builder.annotations.Buildable;
 import io.sundr.builder.annotations.BuildableReference;
 import org.jsonschema2pojo.GenerationConfig;
 
 public class KubernetesTypeAnnotator extends KubernetesCoreTypeAnnotator {
 
-    private static final String BUILDABLE_REFERENCE_VALUE = "value";
+  public static final String BUILDABLE_REFERENCE_VALUE = "value";
 
   public KubernetesTypeAnnotator(GenerationConfig generationConfig) {
     super(generationConfig);
   }
 
   @Override
-    public void processBuildable(JDefinedClass clazz) {
-      try {
-        JAnnotationUse buildable = clazz.annotate(Buildable.class)
+  public void processBuildable(JDefinedClass clazz) {
+    try {
+      JAnnotationUse buildable = clazz.annotate(Buildable.class)
           .param("editableEnabled", false)
           .param("validationEnabled", false)
           .param("generateBuilderPackage", false)
           .param("lazyCollectionInitEnabled", false)
           .param("builderPackage", "io.fabric8.kubernetes.api.builder");
 
-        JAnnotationArrayMember arrayMember = buildable.paramArray("refs");
-        arrayMember.annotate(BuildableReference.class).param(BUILDABLE_REFERENCE_VALUE, new JCodeModel()._class("io.fabric8.kubernetes.api.model.ObjectMeta"));
-        arrayMember.annotate(BuildableReference.class).param(BUILDABLE_REFERENCE_VALUE, new JCodeModel()._class("io.fabric8.kubernetes.api.model.LabelSelector"));
-        arrayMember.annotate(BuildableReference.class).param(BUILDABLE_REFERENCE_VALUE, new JCodeModel()._class("io.fabric8.kubernetes.api.model.Container"));
-        arrayMember.annotate(BuildableReference.class).param(BUILDABLE_REFERENCE_VALUE, new JCodeModel()._class("io.fabric8.kubernetes.api.model.PodTemplateSpec"));
-        arrayMember.annotate(BuildableReference.class).param(BUILDABLE_REFERENCE_VALUE, new JCodeModel()._class("io.fabric8.kubernetes.api.model.ResourceRequirements"));
-        arrayMember.annotate(BuildableReference.class).param(BUILDABLE_REFERENCE_VALUE, new JCodeModel()._class("io.fabric8.kubernetes.api.model.IntOrString"));
-        arrayMember.annotate(BuildableReference.class).param(BUILDABLE_REFERENCE_VALUE, new JCodeModel()._class("io.fabric8.kubernetes.api.model.ObjectReference"));
-        arrayMember.annotate(BuildableReference.class).param(BUILDABLE_REFERENCE_VALUE, new JCodeModel()._class("io.fabric8.kubernetes.api.model.LocalObjectReference"));
-        arrayMember.annotate(BuildableReference.class).param(BUILDABLE_REFERENCE_VALUE, new JCodeModel()._class("io.fabric8.kubernetes.api.model.PersistentVolumeClaim"));
-
-      } catch (JClassAlreadyExistsException e) {
-        e.printStackTrace();
-      }
+      JAnnotationArrayMember arrayMember = buildable.paramArray("refs");
+      arrayMember.annotate(BuildableReference.class).param(BUILDABLE_REFERENCE_VALUE,
+          new JCodeModel()._class("io.fabric8.kubernetes.api.model.ObjectMeta"));
+      arrayMember.annotate(BuildableReference.class).param(BUILDABLE_REFERENCE_VALUE,
+          new JCodeModel()._class("io.fabric8.kubernetes.api.model.LabelSelector"));
+      arrayMember.annotate(BuildableReference.class).param(BUILDABLE_REFERENCE_VALUE,
+          new JCodeModel()._class("io.fabric8.kubernetes.api.model.Container"));
+      arrayMember.annotate(BuildableReference.class).param(BUILDABLE_REFERENCE_VALUE,
+          new JCodeModel()._class("io.fabric8.kubernetes.api.model.PodTemplateSpec"));
+      arrayMember.annotate(BuildableReference.class).param(BUILDABLE_REFERENCE_VALUE,
+          new JCodeModel()._class("io.fabric8.kubernetes.api.model.ResourceRequirements"));
+      arrayMember.annotate(BuildableReference.class).param(BUILDABLE_REFERENCE_VALUE,
+          new JCodeModel()._class("io.fabric8.kubernetes.api.model.IntOrString"));
+      arrayMember.annotate(BuildableReference.class).param(BUILDABLE_REFERENCE_VALUE,
+          new JCodeModel()._class("io.fabric8.kubernetes.api.model.ObjectReference"));
+      arrayMember.annotate(BuildableReference.class).param(BUILDABLE_REFERENCE_VALUE,
+          new JCodeModel()._class("io.fabric8.kubernetes.api.model.LocalObjectReference"));
+      arrayMember.annotate(BuildableReference.class).param(BUILDABLE_REFERENCE_VALUE,
+          new JCodeModel()._class("io.fabric8.kubernetes.api.model.PersistentVolumeClaim"));
+      addBuildableTypes(arrayMember);
+    } catch (JClassAlreadyExistsException e) {
+      e.printStackTrace();
     }
+  }
+
+  protected void addBuildableTypes(JAnnotationArrayMember arrayMember) throws JClassAlreadyExistsException {
+
+  }
 
 }